Non posix mode does not work for ghostscript. If a filename has a quote surrounding it, gs does not strip it out. Since " is invalid for filenames on windows, this does not work.
I've configured it to only parse once over the input string, thus avoiding the original reason for switching to Non posix mode.

Also, it is set up to allow setting of posix mode manually if the user needs to.

Non posix mode still handles quotes slightly oddly, in that if a command such as -I"c:/program files/gs/gs9.05/fonts" is input, two tokens are returned.

This can be fixed, if necessary, in Non posix mode since tokens are examined individually, instead of just using shlex.split() and getting a list.